Update get-bam-file-from-directory to match wiki
Update get-bam-file-from-directory changes to match cwl-ica wiki with move from rabix/cwl-ts repo to cwl-ts-auto
