+++ expressions/get-bam-file-from-directory/1.0.1/typescript-expressions/jest.config.js
@@ -1,5 +1,10 @@
-/** @type {import('ts-jest/dist/types').InitialOptionsTsJest} */
+/** @type {import(\'ts-jest/dist/types\').InitialOptionsTsJest} */
 module.exports = {
   preset: 'ts-jest',
   testEnvironment: 'node',
-};
+  testRegex: "(tests/.*|(\.|/)(test|spec))\.(ts|js)x?$",
+  collectCoverage: true,
+  coverageReporters: [
+    "text-summary"
+  ]
+}
