+++ expressions/get-bam-file-from-directory/1.0.1/typescript-expressions/package.json
@@ -1,5 +1,12 @@
 {
-  "scripts": {
-    "test": "jest"
+  "name": "get-bam-file-from-directory__1.0.1",
+  "packageManager": "yarn@3.2.3",
+  "devDependencies": {
+    "@types/jest": "^29.0.3",
+    "@types/node": "^18.7.23",
+    "cwl-ts-auto": "^0.1.3",
+    "jest": "^29.0.3",
+    "ts-jest": "^29.0.2",
+    "typescript": "^4.8.3"
   }
 }
